ChengHsin/line-ai-bot/requirements.txt
Flask==2.3.3
line-bot-sdk==3.5.0
openai==0.28.1
python-dotenv==1.0.0
gunicorn==21.2.0